refactored the utilities functions into a transformer and vetorizer file as well as adding some rough notes to how we'd like our docmap and topicmap classes to work.
