+++ textmap/vectorizers.py
@@ -1,0 +1,25 @@
+
+class WordVectorizer (BaseEstimator, TransformerMixin):
+    def __init__(self):
+        # make sure we pass the before/after as a switch
+        pass
+
+    def fit(self):
+        # use tokenizers to build list of lists
+        # co-occurence vectorizer
+        # information transformer
+        # em transformer
+        pass
+    pass
+
+class DocVectorizer (BaseEstimator, TransformerMixin):
+    def __init__(self):
+        pass
+
+    def fit(self):
+        # use tokenizers to build list of lists
+        # NgramVectorizer
+        # information transformer
+        # Em transformer
+        pass
+    pass
