+++ textmap/wordmap.py
@@ -3,5 +3,9 @@ from sklearn.base import BaseEstimator, ClassifierMixin, TransformerMixin
 from sklearn.utils.validation import check_X_y, check_array, check_is_fitted
 
 class WordMAP (BaseEstimator, TransformerMixin):
+    def __init__(self):
 
+    def fit(self):
+        # use tokenizers to build list of lists
+        count_matrix =
     pass
